Update README.md
# Hellow, I am Aditya-Kumar, Master's student in DU
# Theoretical Condensed Matter Computation

Welcome to the repository for self-made Python, MATLAB, Scilab, and Mathematica code for computational analysis in theoretical condensed matter physics.

## Introduction

This repository is open source and designed to aid in understanding and learning theoretical condensed matter concepts. The code provided covers a variety of topics, including but not limited to:

1. **Complex Hamiltonian in Theoretical Condensed Matter**
   - Berry Phase and Berry Connection calculations
   - Wannier Center, Dirac Cone, etc.
   - Density of States, Local Density of States, Chern Number, etc.
   - Graphene structure, 0D, 1D, 2D, 3D SSH model, Rice-Mele model
   - Charge Pumping, Z2 Invariance, Band Gap, Band Energy, etc.
   - Any nearest hopping calculations

## Usage

Feel free to explore the code and use it for your own projects, research, or educational purposes. Contributions and improvements are welcome!

## Code Structure

The repository is organized into folders corresponding to different theoretical concepts. Each folder contains code written in Python, MATLAB, Scilab, or Mathematica, providing flexibility for different users.

## How to Contribute

If you find issues, have suggestions, or want to contribute new features, please open an issue or submit a pull request. Your contributions make this repository a valuable resource for the community.

## License

This project is under no License, allowing you to use the code for various purposes.

Happy coding!